タイムラインのViewとモデルの紐付け処理を外だしした
